# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully"
          }
        },
        "summary": "Register a new customer",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns access token and user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          }
        },
        "summary": "Login and get JWT",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Logged out successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Logout (invalidates token)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "AuthController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Current user data"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user profile",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "If email exists, reset link sent"
          }
        },
        "summary": "Request password reset (emits event → Notifications sends email)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successfully"
          }
        },
        "summary": "Reset password with token from email",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/cars": {
      "get": {
        "operationId": "CarsController_getAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "make",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "model",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "yearMin",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "yearMax",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "priceMin",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "priceMax",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "slug",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of cars"
          }
        },
        "summary": "List cars (paginated, with filters)",
        "tags": [
          "Cars"
        ]
      },
      "post": {
        "operationId": "CarsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCarDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Car created"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create car (emits car.created event)",
        "tags": [
          "Cars"
        ]
      }
    },
...
    "/api/cars/slug/{slug}": {
      "get": {
        "operationId": "CarsController_getBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Car details"
          }
        },
        "summary": "Get car by slug",
        "tags": [
          "Cars"
        ]
      }
    },
    "/api/cars/{id}": {
      "get": {
        "operationId": "CarsController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Car details"
          }
        },
        "summary": "Get car by ID",
        "tags": [
          "Cars"
        ]
      },
      "put": {
        "operationId": "CarsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCarDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Car updated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update car (emits car.updated event)",
        "tags": [
          "Cars"
        ]
      },
      "delete": {
        "operationId": "CarsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Car deleted"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete car (emits car.deleted event)",
        "tags": [
          "Cars"
        ]
      }
    },
    "/api/cars/{id}/image": {
      "post": {
        "operationId": "CarsController_uploadImage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Image uploaded"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upload car image",
        "tags": [
          "Cars"
        ]
      }
    },
...
    "/api/admin/dashboard": {
      "get": {
        "operationId": "AdminController_getDashboard",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Dashboard stats (cars, news, test drives, contacts)"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get admin dashboard (aggregates event-driven stats)",
        "tags": [
          "Admin"
        ]
      }
    }
  },
  "info": {
    "title": "Baccouche Motors API Gateway",
    "description": "API Gateway for Baccouche Motors microservices.\n\n## Event-Driven Design (RabbitMQ)\nThis system follows a strict **Event-Driven Design** pattern:\n- **Domain events** are emitted on create/update/delete operations\n- **Consumers** react asynchronously (emails, stats, audit)\n\n| Event | Producer | Consumers |\n|-------|----------|-----------|\n| `auth.forgot-password-requested` | Auth | Notifications (email) |\n| `car.created` / `updated` / `deleted` | Cars | Admin (stats) |\n| `news.created` / `updated` / `deleted` | News | Admin (stats) |\n| `contact.created` / `updated` | Contacts | Notifications, Admin |\n| `test-drive.created` / `status-changed` | Test Drives | Notifications, Admin |\n\n**Flow**: Gateway → RabbitMQ RPC (sync) → Microservice → emits to `baccouche.events` exchange → Consumers (Notifications, Admin)",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "AuthRegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 8
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password",
          "firstName",
          "lastName",
          "phone"
        ]
      },
      "LoginUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "user": {
            "type": "object",
            "description": "User object"
          },
          "auth_token": {
            "type": "string",
            "description": "JWT access token"
          }
        },
        "required": [
          "success",
          "user",
          "auth_token"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Reset token from email"
          },
          "newPassword": {
            "type": "string",
            "minLength": 8
          }
        },
        "required": [
          "token",
          "newPassword"
        ]
      },
...
      "CreateCarDto": {
        "type": "object",
        "properties": {
          "make": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "year": {
            "type": "number"
          },
          "price": {
            "type": "number"
          },
          "slug": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "specs": {
            "type": "object"
          },
          "vin": {
            "type": "string"
          },
          "mileage": {
            "type": "number"
          },
          "color": {
            "type": "string"
          },
          "fuelType": {
            "type": "string"
          },
          "transmission": {
            "type": "string"
          }
        },
        "required": [
          "make",
          "model",
          "year",
          "price"
        ]
      },
      "UpdateCarDto": {
        "type": "object",
        "properties": {
          "make": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "year": {
            "type": "number"
          },
          "price": {
            "type": "number"
          },
          "slug": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "specs": {
            "type": "object"
          },
          "vin": {
            "type": "string"
          },
          "mileage": {
            "type": "number"
          },
          "color": {
            "type": "string"
          },
          "fuelType": {
            "type": "string"
          },
          "transmission": {
            "type": "string"
          }
        }
      },
...
    }
  }
}